+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Added trigger ```scrolling.uk.document``` after update grid layout
   - Fixed sticky wrapper resize on window resize/orientation change
   - Fixed nestable+sortable: serialize only includes data attributes from HTML
+  - Fixed video mute bug in slideshow for vimeo player
 
 
 ### 2.17.0 (February 20, 2015)